+++ src/app/creator-builds/page.tsx
@@ -1,17 +1,11 @@
-'use client'
-
-import { useSearchParams } from 'next/navigation'
+'use server'
 
 import { FeaturedBuilds } from '@/app/creator-builds/FeaturedBuilds'
 import { BuildListFilters } from '@/features/filters/components/BuildListFilters'
-import { parseBuildListFilters } from '@/features/filters/lib/parseBuildListFilters'
 import { DEFAULT_ITEMS_PER_PAGE } from '@/features/pagination/constants'
 import { PageHeader } from '@/features/ui/PageHeader'
 
-export default function Page() {
-  const searchParams = useSearchParams()
-  const buildListFilters = parseBuildListFilters(searchParams)
-
+export default async function Page() {
   return (
     <>
       <PageHeader
@@ -20,19 +14,11 @@ export default function Page() {
       />
 
       <div className="mb-8 flex w-full max-w-3xl items-center justify-center">
-        <BuildListFilters
-          filters={buildListFilters}
-          key="featured-build-filters"
-        />
+        <BuildListFilters key="featured-build-filters" />
       </div>
-      {buildListFilters && (
-        <div className="mb-4 grid w-full grid-cols-1 gap-2">
-          <FeaturedBuilds
-            buildListFilters={buildListFilters}
-            itemsPerPage={DEFAULT_ITEMS_PER_PAGE}
-          />
-        </div>
-      )}
+      <div className="mb-4 grid w-full grid-cols-1 gap-2">
+        <FeaturedBuilds itemsPerPage={DEFAULT_ITEMS_PER_PAGE} />
+      </div>
     </>
   )
 }