+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-03-06
+
+### Changed
+
+- The "Builds By Collection" feature has been removed from site. The page link is still active with a message explaining the change.
+
 ## 2024-03-05
 
 ### Added